text,vg/draw: move TextHandler and TextStyle to text
