+++ text/latex.go
@@ -16,7 +16,6 @@ import (
 	"github.com/go-latex/latex/tex"
 
 	"gonum.org/v1/plot/vg"
-	"gonum.org/v1/plot/vg/draw"
 )
 
 // Latex parses, formats and renders LaTeX.
@@ -26,7 +25,7 @@ type Latex struct {
 	DPI float64
 }
 
-var _ draw.TextHandler = (*Latex)(nil)
+var _ Handler = (*Latex)(nil)
 
 // Lines splits a given block of text into separate lines.
 func (hdlr Latex) Lines(txt string) []string {
@@ -74,7 +73,7 @@ func (hdlr Latex) Box(txt string, fnt vg.Font) (width, height, depth vg.Length) 
 
 // Draw renders the given text with the provided style and position
 // on the canvas.
-func (hdlr Latex) Draw(c vg.Canvas, txt string, sty draw.TextStyle, pt vg.Point) {
+func (hdlr Latex) Draw(c vg.Canvas, txt string, sty TextStyle, pt vg.Point) {
 	cnv := drawtex.New()
 	fnts := &ttf.Fonts{
 		Rm:      sty.Font.Font(),
@@ -137,7 +136,7 @@ func (hdlr Latex) dpi() float64 {
 
 type latex struct {
 	cnv vg.Canvas
-	sty draw.TextStyle
+	sty TextStyle
 	pt  vg.Point
 
 	w vg.Length
